some missed renamed imports
