+++ packages/react-reconciler/src/forks/ReactFiberErrorDialog.native.js
@@ -7,7 +7,7 @@
  * @flow
  */
 
-import type {Fiber} from '../ReactFiber.new';
+import type {Fiber} from '../ReactFiber.old';
 import type {CapturedValue} from '../ReactCapturedValue';
 
 import {ClassComponent} from '../ReactWorkTags';
